Transaction add functionality implemented

+++ app/src/main/java/com/example/moneyflow/data/Transaction.kt
@@ -28,7 +28,7 @@ data class Transaction(
     val categoryId: Int,
     val walletId: Int,
 
-    val sum: Int,
+    val sum: Double,
     val isIncome: Boolean,
 
     val note: String?,
